Preliminary EDA by ERG
Just a start... We probably want to rename variables! But the pre- / post-pandemic question make sense to me. There also seems to be a seasonal variation in rates of collision as Sairam noted-- that might be harder to explore (in past have done this using month as a factor, but functional fitting would make more sense).