Borough level Analysis and modelling
This Python file contains the Borough-level analysis, the EDA, and the modeling. This Python file was run on google collab, so please replace the file path with your local file path. The pycaret library has been utilized for the modeling, so you can install it using "pip install pycaret" to download the latest version.